refactor(core): rearrange packages and reduce core package dependency size
Summary:
Move logic and tests from the core package to the other dependent packages so that the core package doesn't need to use the browser installed with puppeteer.

The main reason behind all this is that if someone just want to use `memlab/core` in their code (without the other memlab packages), they can choose to not download Chromium when doing `npm install memlab/core`.

Differential Revision: D37063223

fbshipit-source-id: e856e0a0fc24a9d9cf7663f91ada69e7b97f09df