doc: tweak the terminal replay demo
Summary:
This diff tweaks the speed of terminal replay on memlab doc site home page.

Some parts of the demo is waiting for too long (e.g., npm install), users may think that the demo gets stuck and leave. While some parts of the demo proceeds to the next command too quickly (users may not be able to read the terminal results).

Reviewed By: tulga1970

Differential Revision: D37834868

fbshipit-source-id: c6f24d34e13382c8856399a820303fe1aa9193da