fix(core): optimize trace JSON size
Summary:
When `--trace-all-object` is enabled, some clustering may get a OOM crash in node.js (found by Tulga).

After investigation, I realize the `JSONify` path may generate trace JSON that is too big (e.g., 50MB), and if we cache all of them in memory (up to 5k traces), it will use up the JS heap and we get an OOM in memlab (lol).

This diff optimizes the trace JSON size by limiting the number of properties encoded in the trace JSON file.

Reviewed By: tulga1970

Differential Revision: D38220389

fbshipit-source-id: 28d1421de38f9cf28c46d5f08b9d3b2d57f76f47