feat(cli): integrate heap analysis with the heap viewer
Summary:
By default, the Object box now shows a list of memory leaks or chosen objects.

After this change, MemLab will put most objects related to memory issues in the `Objects` box, so users can view and interact with the information related to those objects in the other UI boxes.

Differential Revision: D38997388

fbshipit-source-id: b70ebd610334b0c1e8d23804509a6b6c5c99e3ac