feat: de-emphasize less useful edges for debugging
Summary:
This diff de-emphasize heap objects and references that are less useful for debugging, so that people can pay more attention to those with more signals.

The usefulness of those heap objects and references are determined by the type of those entities. For example, internal references or code heap objects are less useful for debugging since they are generated by browser (often not a manifestation of leaks in JS code from website).

Differential Revision: D39162670

fbshipit-source-id: 2cff955f3567eb3aa8ea4f7d033dd3a6961215c9