chore: force sample retainer traces to objects with non-trivial self size
Summary: To ensure that the leak trace clusters we build from samples represent heap objects of significant size, we should prioritize selecting leak traces that contain heap objects with a self size of at least 100KB when sampling. This way, we can be sure that the clusters won't miss heap objects with non-trivial self sizes.

Reviewed By: tulga1970

Differential Revision: D42134368

fbshipit-source-id: 4986b5a2ff41b8c4076ea57a1c83d18699823e13