chore(core): optimize leak trace sampling for clustering and trace diffing
Summary:
This change improves the sampling algorithm for clustering leak traces by filtering out traces that are longer than the 80th percentile of the entire set. This prevents the generation of clusters that contain excessively long trace samples and speeds up the heuristic clustering process.

Additionally, the heuristic clustering module now displays the current progress, such as "clustering trace: ${number}/${totalSize}", which is useful when clustering or diffing large sets of traces.

Reviewed By: tulga1970

Differential Revision: D42592259

fbshipit-source-id: 5e6bb8a4d4fbae0794e666401c445371828dc986