feat: memlab diff-leaks logic can accept multiple control working directories
Summary:
This change enables MemLab diff-leaks logic to accept multiple control working directories.

**Context:** The web app under test may be non-deterministic; and, therefore a trace absent in control heap snapshot doesn't necessarily mean it will never appear in the control heap snapshots from other control runs.

To reduce this kind of non-determinism, this change enables MemLab's leak trace diffing algorithm to be able to accept heap snapshot from multiple control runs. If a leak trace cluster appears in any of the control heap snapshots, it is considered as not triggered uniquely in the treatment group.

Reviewed By: tulga1970

Differential Revision: D42949624

fbshipit-source-id: a3bf69f247ab6ca1bc635f79ac6735dfad453730