docs: Update example usage of analyze API in documentation (#73)
Summary:
This PR updates the example usage of the analyze API in the documentation for the memlab/api package. The existing example was missing the analyze import and was not functional. This PR fixes the example by adding the necessary import and function call.

The existing example was incorrect and could cause confusion for users trying to use the analyze API. This change ensures that the example is correct and functional, making it easier for users to understand and use the API.

Pull Request resolved: https://github.com/facebook/memlab/pull/73

Reviewed By: tulga1970

Differential Revision: D45201444

Pulled By: JacksonGL

fbshipit-source-id: 40cb9be41ea79460bb8c9fcf5b448e86e8d3bca6