feat(cli): add CLI options to allow specifying filter traces by object and reference names
Summary:
This change adds the new `--trace-contains` flag to specify which substring that must be included in leak traces.

For example, the following command will only filter out and compare leak traces that contains the `abc` substring in any one of the heap object name or reference name.
```
memlab diff-leaks --trace-contains abc
```

So the following example traces will be included in leak trace diffing:

```
window --abcde--> [object] --> ...
window --varName--> [xyzabcde] --> ...
```

This can be useful for targeting leak traces containing specific object or reference and reducing the sample size when used with the `--trace-all-objects` flag, which often leads to a huge sample size (over 100k+ traces)

Reviewed By: tulga1970

Differential Revision: D45261183

fbshipit-source-id: e70378e61cea84ea906e5919ecbae88c94ffdc1f