feat(e2e): support filtering references for retainer traces (#106)
Summary:
This diff adds a new callback that can be used to define a logic to decide whether a reference should be considered as part of the retainer trace. The callback is called for every reference (edge) in the heap snapshot.

The callback accepts the following parameters
  * `edge` - the reference (edge) that is considered for calcualting the retainer trace
  * `snapshot` - the snapshot of target interaction
  * `isReferenceUsedByDefault` - MemLab has its own default logic for whether a reference should be considered as part of the retainer trace, if this parameter is true, it means MemLab will consider this reference when calculating the retainer trace.

The return value indicates whether the given reference should be considered when calculating the retainer trace. Note that when this callback returns true, the reference will only be considered as a candidate for retainer trace, so it may or may not be included in the retainer trace; however, if this callback returns false, the reference will be excluded.

Note that by excluding a dominator reference of an object (i.e., an edge that must be traveled through to reach the heap object from GC roots), the object will be considered as unreachable in the heap graph; and therefore, the reference and heap object will not be included in the retainer trace detection and retainer size calculation.

```
// save as leak-filter.js
module.exports = {
  retainerReferenceFilter(edge, _snapshot, _leakedNodeIds) {
    // exclude react fiber references
    if (edge.name_or_index.toString().startsWith('__reactFiber$')) {
      return false;
    }
    return true;
  }
};
```

Use the leak filter definition in command line interface:
```
memlab find-leaks --leak-filter <PATH TO leak-filter.js>
```

```
memlab run --scenario <SCENARIO FILE> --leak-filter <PATH TO leak-filter.js>
```

Differential Revision: D53167155

fbshipit-source-id: b9006fd309276452acbe216c577f34c28b071ffa