fix(core): support parsing heap snapshots with over 100 million nodes (#125)
Summary:
Fixing an issue reported by OSS user:
https://github.com/facebook/memlab/issues/125

The heap snapshot parser is using an object literal as dictionary to map from node id to node index. Currently the JS engines does not support storing more than 180 million KV pairs in objects or Map. The solution is to replace the object literal with a two-level indirect map that shards the node id. So in theory the map can now handle (100 million) ^ 2 nodes, which should no longer be a bottleneck.

Reviewed By: twobassdrum

Differential Revision: D61445058

fbshipit-source-id: b7f6f230ab9f673b513cd496a28ca66f935c9d1c