feat(core): support console annotations (#124)
Summary:
When dev set the following config, we want stack traces to be printed even if `lowLevel` output is muted (stack traces is output at `lowLevel` for various reasons such as coloring, but semantically it is part of the error output).
```
memlabConfig.muteConfig = {
      muteError: false,
      muteWarning: false,
      muteInfo: true,
      muteSuccess: true,
      muteLog: true,
      muteTable: true,
      muteTrace: true,
      muteTopLevel: true,
      muteHighLevel: true,
      muteMidLevel: true,
      muteLowLevel: true,
};
memlabConfig.verbose = true;
```

This diff patches the MemLab console output module so that stack traces is still printed when `muteLowLevel` is `true` while `muteError` is `false`.

Differential Revision: D63012111

fbshipit-source-id: 535bde1a3e7287a2d5b33c75f48b04f708a44be8