feat(lens): supports pseudo virtualization of leak outline overlay
Summary: This diff enables pseudo virtualization in Memory Len's memory leak outline overlay so that when the outline is scrolled out of the viewport it is hidden, while when the ouline is scrolled into the viewport it is displayed. This may help reduce the rendering overhead added by the outlines.

Reviewed By: twobassdrum

Differential Revision: D73407573

fbshipit-source-id: c2547d66968c7a1500681269cf374d35fe473d2a