chore(lens): debounce react component stack panel update to avoid UI flickering
Summary: This diff debounces updates to the React component stack panel to prevent a flickering loop caused by rapid mouse transitions. The issue occurs when the mouse leaves leak outline A and enters the React component stack, which causes the stack to disappear due to the outline change. If leak outline B is underneath the stack, it triggers a re-render of the stack. This causes the mouse out to be out of B (because the react component stack is rendered on top of it) and back into the react component stack panel, this causes the same <leaving outline, enter component stack, component stack disappear, re-enter outline> behavior to repeat, resulting in an infinite loop of UI flickering.

Reviewed By: twobassdrum

Differential Revision: D73486515

fbshipit-source-id: 523c5b0f821c1c60f774270ab26d988732cc6a0b